We did a lot read Description
James's shop system, with all the UI and the placement system is working, the currency was fixed, the day/night cycle, UI was added for everything. Created loading screen to allow world to be created, hooked up the main menu with the game, made a pause system that pauses evrythign in the game, pause panel, that can resume and main menu, optimized the camera.